+++ src/postgres/tree/PostgresStoredProceduresTreeItem.ts
@@ -6,8 +6,7 @@
 import { ClientConfig } from 'pg';
 import { TreeItemIconPath } from "vscode-azureextensionui";
 import { getThemedIconPath } from "../../constants";
-import { getPostgresProcedureQueryRows } from '../getPostgresProcedureQueryRows';
-import { IPostgresProceduresQueryRow } from '../IPostgresProceduresQueryRow';
+import { getPostgresProcedureQueryRows, IPostgresProceduresQueryRow } from '../getPostgresProcedureQueryRows';
 import { PostgresDatabaseTreeItem } from './PostgresDatabaseTreeItem';
 import { PostgresResourcesTreeItemBase } from './PostgresResourcesTreeItemBase';
 import { PostgresStoredProcedureTreeItem } from './PostgresStoredProcedureTreeItem';
@@ -32,22 +31,7 @@ export class PostgresStoredProceduresTreeItem extends PostgresResourcesTreeItemB
     }
 
     public async loadMoreChildrenImpl(): Promise<PostgresStoredProcedureTreeItem[]> {
-        // Adapted from https://aka.ms/AA83fg8
-        const storedProceduresQuery: string = `select n.nspname as schema,
-            p.proname as name,
-            p.oid as oid,
-            pg_get_function_arguments(p.oid) as args,
-            case when l.lanname = 'internal' then p.prosrc
-                else pg_get_functiondef(p.oid)
-                end as definition
-            from pg_proc p
-            left join pg_namespace n on p.pronamespace = n.oid
-            left join pg_language l on p.prolang = l.oid
-            where n.nspname not in ('pg_catalog', 'information_schema')
-                and p.prokind = 'p'
-            order by name;`;
-
-        const rows: IPostgresProceduresQueryRow[] = await getPostgresProcedureQueryRows(this, storedProceduresQuery);
+        const rows: IPostgresProceduresQueryRow[] = await getPostgresProcedureQueryRows(this);
         return rows.map(row => new PostgresStoredProcedureTreeItem(
             this,
             row,
